GridSearchService refactor (#140)
* Binder update and npm publish (#139)

* Refactored GridSearchService into another file, added some comments
