+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jupyterlab-tabular-data-editor",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "A tool for editing csv files natively in JupyterLab",
   "keywords": [
     "jupyter",
